File populate
Added Jade and corresponding compiled html files as well as Sass/CSS
and images. For some reason I have yet to figure out, the markup is
indented weird for the mixins when I compile from Jade. Anyone reading
this crap should keep in mind that I haven’t coded with Jade ever
before, and haven’t written html or css for many years. So picture me
as kind of a noob. Any feedback would be welcome and appreciated. I’m
also a total noob to Git, so this is all pretty much an experiment.